backend/pyproject.toml
[project]
name = "telegram-media-dl"
version = "0.1.0"
description = "Telegram media browser and batch downloader with Web UI, Bot, and Mini App frontends"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "telethon>=1.36.0",
    "python-telegram-bot>=21.0",
    "cryptography>=43.0.0",
    "aiolimiter>=1.1.0",
    "imageio-ffmpeg>=0.5.1",
    "pyyaml>=6.0.2",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "pillow>=11.0.0",
    "qrcode[pil]>=8.0",
    "python-dotenv>=1.2.2",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP", "SIM", "RUF"]
ignore = ["E501", "SIM105", "RUF006", "RUF012", "RUF059", "B023", "UP042"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
